package: update dependencies and regenerate lock file
due to a bug in npm 7, when upgrading the lockfile from v1 to v2, the
`hasInstallScript` key was missing from the `esbuild`, which meant it
didn’t install correctly in some cases. A fix for this is to regenerate
the lockfile and install all dependencies from their package.json definition,
which upgrades the upgradable ones.

More detailed explanation (comment by the esbuild author):
https://github.com/npm/cli/issues/2606#issuecomment-819381036